Activitys moved to new packages
