+++ app/src/main/java/com/example/koffer/MainActivity.java
@@ -4,6 +4,8 @@ import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
+import com.example.koffer.view.SlideActivity;
+
 public class MainActivity extends AppCompatActivity {
 
     @Override
