+++ app/src/main/java/com/example/koffer/view/RegisterActivity.java
@@ -11,9 +11,7 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
-import com.example.koffer.MainActivity;
 import com.example.koffer.R;
-import com.example.koffer.SelectLoginActivity;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
