+++ app/src/main/java/com/example/koffer/view/SelectLoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.koffer;
+package com.example.koffer.view;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -6,8 +6,7 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 
-import com.example.koffer.view.LoginActivity;
-import com.example.koffer.view.RegisterActivity;
+import com.example.koffer.R;
 
 public class SelectLoginActivity extends AppCompatActivity {
