+++ app/src/main/res/layout/activity_select_login.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SelectLoginActivity"
+    tools:context=".view.SelectLoginActivity"
     android:orientation="vertical"
     android:background="@color/colorPrimary"
     android:padding="25dp">
